pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = ""
authors = ["ph <robotnaoborot@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django_rest_framework = "latest"
drf-jsonschema = {git = "https://github.com/pawnhearts/drf_jsonschema"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
